+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
   [#81](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/81)
 * `db backup` CLI command and `dump` command in app db client running `pg_dump`
   [#93](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/93)
+* `data backup` CLI command and app data client using GDAL
+  [#108](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/108)
 
 ### Changed